refactor(api-go): flat-events swap — remotes/membership cluster onto wire bindings
ChannelRemoteDeleted / ChannelRemotesSynced / ChannelMessagesSynced /
ChannelMembership{Added,Removed} payloads become 'type X = wire.X' aliases;
their hand-rolled envelopes in internal/shared/events are deleted in the same
commit. Publishers construct types.NewIntegrationEvent(wire.<X>EventName, …).
Disclosed type adaptations only: summary counts cast int→int32 at the two
whatsmeow construction sites (+1 test fixture) — the binding types them int32;
marshaled bytes identical.

remote_created / remote_updated stay hand-rolled (BLOCKED: RemoteType
USER→ContactKind CONTACT value-set divergence — see the classification table in
the delivery/seen/presence swap commit).

Wire-identity: 22 goldens byte-identical (goldens untouched); gateway
openapi.json regen byte-identical; go build/vet/test 12 pkgs ok.

+++ packages/api/go/internal/channel/services/gateway/whatsapp/bootstrap_test.go
@@ -304,8 +304,8 @@ func TestApplyHistorySyncBatch_EmitsSummaryOnlyOnInsert(t *testing.T) {
 			summaryEvt := ctxevents.NewMessagesSyncedEvent(ch.instanceID, ch.ownerID, ctxevents.ChannelMessagesSyncedPayload{
 				ChannelID: ch.instanceID,
 				OwnerID:   ch.ownerID,
-				Total:     len(records),
-				Inserted:  insertedCount,
+				Total:     int32(len(records)),
+				Inserted:  int32(insertedCount),
 			})
 			_ = domainRepo.Save(context.Background(), summaryEvt)
 		}
@@ -337,8 +337,8 @@ func TestApplyHistorySyncBatch_EmitsSummaryOnlyOnInsert(t *testing.T) {
 			summaryEvt := ctxevents.NewMessagesSyncedEvent(ch.instanceID, ch.ownerID, ctxevents.ChannelMessagesSyncedPayload{
 				ChannelID: ch.instanceID,
 				OwnerID:   ch.ownerID,
-				Total:     len(records),
-				Inserted:  insertedCount,
+				Total:     int32(len(records)),
+				Inserted:  int32(insertedCount),
 			})
 			_ = domainRepo.Save(context.Background(), summaryEvt)
 		}
